Refactor and add type to features table
